imagesgs

+++ src/main/java/blue/thejester/tinkerprofiler/item/TraitProfiler.java
@@ -5,11 +5,15 @@ import c4.conarm.lib.materials.ArmorMaterialType;
 import c4.conarm.lib.materials.CoreMaterialStats;
 import c4.conarm.lib.materials.PlatesMaterialStats;
 import c4.conarm.lib.materials.TrimMaterialStats;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.block.model.BakedQuad;
+import net.minecraft.client.renderer.block.model.IBakedModel;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.EnumActionResult;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
@@ -47,6 +51,9 @@ public class TraitProfiler extends Item {
 
     @Override
     public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
+        if(worldIn.isRemote) {
+            return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, playerIn.getHeldItem(handIn));
+        }
         File outFile = new File("tinkers-data.log");
         if (outFile.exists()) {
             outFile.delete();
@@ -130,6 +137,7 @@ public class TraitProfiler extends Item {
                 }
             }
             row.configureTraits(mat);
+            row.setImage(mat);
             mats.add(row);
         }
         return String.format("{\"materials\": [%s], \"traits\": %s}",
